requirements.txt
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
Pillow==7.2.0
psycopg2==2.8.5
